Fixed bug in eviction queue management
Turns out we forgot to also remove a frame from the eviction queue when
it’s marked as free.

Without this fix, it would seem that the eviction queue would continue
to become very large and full of duplicate entries, since frames are
added to the back of the queue every time they’re populated.